requirements.txt
accelerate==1.10.1
aiofiles==25.1.0
alembic==1.17.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
argcomplete==3.6.3
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
autopage==0.5.2
bcrypt==5.0.0
beautifulsoup4==4.14.2
boto3==1.40.56
botocore==1.40.56
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cliff==4.11.0
cmaes==0.12.0
cmd2==2.7.0
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.8.0
executing==2.2.1
fiftyone==1.9.0
fiftyone-brain==0.21.3
fiftyone_db==1.3.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
glob2==0.7
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
humanize==4.14.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lazy_loader==0.4
lia-web==0.2.3
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.9.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime-gpu==1.23.2
opencv-python-headless==4.12.0.88
optuna==2.10.1
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pooch==1.8.2
pprintpp==0.4.0
prettytable==3.16.0
priority==2.0.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.0.0
pybcj==1.0.6
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydash==8.0.5
Pygments==2.19.2
PyMatting==1.1.14
pymongo==4.9.2
PyNaCl==1.6.0
pynndescent==0.5.13
pyparsing==3.2.4
pyperclip==1.11.0
pyppmd==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
pyzstd==0.18.0
rarfile==4.2
referencing==0.37.0
regex==2025.9.1
rembg==2.0.67
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rich-argparse==1.7.1
rpds-py==0.28.0
rtree==1.4.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.42.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.48.0
stevedore==5.5.0
strawberry-graphql==0.284.1
sympy==1.14.0
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.0
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
universal-analytics-python3==1.1.1
urllib3==2.5.0
voxel51-eta==0.15.1
wandb==0.22.2
wcwidth==0.2.13
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.2.0
xmltodict==1.0.2
